\name{adipose}
\docType{data}
\alias{adipose}
\title{Gene Expression in Human White Adipose Tissue}
\description{
            This data set resulted from microarray experiments performed on human white adipose 
            tissue after the separation of its two cellular components: mature adipocytes and stroma 
            vascular fraction (SVF) cells, the non adipose component of the tissue. The purpose of this 
            experimental model was to distinguish the two cellular fractions from a functional perspective 
            and, in the meantime, to help establish the contribution of adipose and non adipose cells in 
            the expression of inflammatory molecules in morbid obesity.
          }
\usage{data(adipose)}
\source{\url{http://corneliu.henegar.info/FunCluster.htm}}
\format{Three data frames containing significant genes specifically expressed in human mature adipocytes or in stroma 
	vascular fraction cells of human white adipose tissue and a reference list with all the genes tested for 
	differential expression during this experiment.}
\references{
            Cancello R, Henegar C, Viguerie N, Taleb S, Poitou C, Rouault C, Coupaye M, Pelloux V, Hugol D, Bouillot JL, 
	    Bouloumie A, Barbatelli G, Cinti S, Svensson PA, Barsh GS, Zucker JD, Basdevant A, Langin D, 
	    Clement K. Reduction of macrophage infiltration and chemoattractant gene expression changes in 
	    white adipose tissue of morbidly obese subjects after surgery-induced weight loss. 
            Diabetes 2005; 54(8):2277-86.
           }
\seealso{
            \code{\link{FunCluster}, \link{annotations}}.
}
\keyword{datasets}