\name{insulin}
\docType{data}
\alias{insulin}
\title{Insulin Coordinated Regulation of Gene Expression in Human Skeletal Muscle}
\description{
            This data set resulted from a study investigating insulin coordinated regulation 
            of gene expression in human skeletal muscle during a 3-hour hyperinsulinemic euglycemic clamp.
          }
\usage{data(insulin)}
\source{\url{http://corneliu.henegar.info/FunCluster.htm}}
\format{Three data frames containing significant up-regulated and down-regulated genes and a reference 
	list with all the genes tested for differential expression during this experiment.}
\references{
            S. Rome, K. Clement, R. Rabasa-Lhoret, E. Loizon, C. Poitou, G. S. Barsh, J. P. Riou,
            M. Laville and H. Vidal. Microarray profiling of human skeletal muscle reveals that 
            insulin regulates approximately 800 genes during a hyperinsulinemic clamp, 
            J Biol Chem 278(20), 18063-8 (2003).
           }
\seealso{
            \code{\link{FunCluster}, \link{annotations}}.
}
\keyword{datasets}