\name{SC.GO.IND.MF.file.annot}
\alias{SC.GO.IND.MF.file.annot}
\docType{data}
\title{Internal routine or variable}
\description{
  This internal routine or variable should not be accessed or modified directly by the user.
}
\seealso{
           \code{\link{FunCluster}, \link{annotations}}.
}
\keyword{internal}
